requirements.txt
Django   == 1.10
markdown == 2.6.6
djangorestframework == 3.5.1
django-rest-swagger == 2.0.6
coreapi == 2.0.8
rest-framework-generic-relations == 1.1.0
